% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{append_entry}
\alias{append_entry}
\title{Append an entry to an audit log}
\usage{
append_entry(log, entry)
}
\arguments{
\item{log}{An \code{audit_log} object.}

\item{entry}{An \code{audit_entry} object.}
}
\value{
The updated \code{audit_log}.
}
\description{
Append an entry to an audit log
}
