% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{as.data.frame.audit_log}
\alias{as.data.frame.audit_log}
\title{Convert audit log to data frame}
\usage{
\method{as.data.frame}{audit_log}(x, ...)
}
\arguments{
\item{x}{An \code{audit_log} object.}

\item{...}{Ignored.}
}
\value{
A \code{data.frame} with columns: \code{timestamp}, \code{event}, \code{block_id},
\code{user}, \code{details}.
}
\description{
Convert audit log to data frame
}
