% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{blockr_ser.audit_log}
\alias{blockr_ser.audit_log}
\alias{blockr_deser.audit_log}
\title{Serialise / deserialise audit log}
\usage{
\method{blockr_ser}{audit_log}(x, ...)

\method{blockr_deser}{audit_log}(x, data, ...)
}
\arguments{
\item{x}{An \code{audit_log} object.}

\item{...}{Ignored.}

\item{data}{Serialised list representation.}
}
\value{
For \code{blockr_ser}: a serialisable list. For \code{blockr_deser}: an
\code{audit_log} object.
}
\description{
Serialise / deserialise audit log
}
